+++ src/main/java/miyucomics/hexpose/actions/item_stack/OpBookSources.kt
@@ -6,9 +6,9 @@ import at.petrak.hexcasting.api.casting.iota.DoubleIota
 import at.petrak.hexcasting.api.casting.iota.Iota
 import at.petrak.hexcasting.api.casting.iota.NullIota
 import miyucomics.hexpose.iotas.DisplayIota
-import miyucomics.hexpose.iotas.getItemStack
 import net.minecraft.item.WrittenBookItem
 import net.minecraft.text.Text
+import ram.talia.moreiotas.api.getItemStack
 
 object OpBookSources : ConstMediaAction {
 	override val argc = 1