+++ src/main/java/miyucomics/hexpose/actions/item_stack/OpGetArmor.kt
@@ -5,13 +5,13 @@ import at.petrak.hexcasting.api.casting.castables.ConstMediaAction
 import at.petrak.hexcasting.api.casting.eval.CastingEnvironment
 import at.petrak.hexcasting.api.casting.getEntity
 import at.petrak.hexcasting.api.casting.iota.Iota
-import miyucomics.hexpose.iotas.ItemStackIota
+import ram.talia.moreiotas.api.casting.iota.ItemStackIota
 
 object OpGetArmor : ConstMediaAction {
 	override val argc = 1
 	override fun execute(args: List<Iota>, env: CastingEnvironment): List<Iota> {
 		val entity = args.getEntity(0, argc)
 		env.assertEntityInRange(entity)
-		return entity.armorItems.map { ItemStackIota.createOptimized(it) }.asActionResult
+		return entity.armorItems.map { ItemStackIota.createFiltered(it) }.asActionResult
 	}
 }