+++ src/main/java/miyucomics/hexpose/actions/item_stack/OpGetHeldStack.kt
@@ -5,9 +5,9 @@ import at.petrak.hexcasting.api.casting.eval.CastingEnvironment
 import at.petrak.hexcasting.api.casting.iota.EntityIota
 import at.petrak.hexcasting.api.casting.iota.Iota
 import at.petrak.hexcasting.api.casting.mishaps.MishapInvalidIota
-import miyucomics.hexpose.iotas.asActionResult
 import net.minecraft.entity.LivingEntity
 import net.minecraft.util.Hand
+import ram.talia.moreiotas.api.asActionResult
 
 class OpGetHeldStack(private var hand: Hand) : ConstMediaAction {
 	override val argc = 1