+++ src/main/java/miyucomics/hexpose/actions/media/OpGetMaxMedia.kt
@@ -13,9 +13,9 @@ import at.petrak.hexcasting.api.casting.iota.NullIota
 import at.petrak.hexcasting.api.casting.iota.Vec3Iota
 import at.petrak.hexcasting.api.misc.MediaConstants
 import at.petrak.hexcasting.xplat.IXplatAbstractions
-import miyucomics.hexpose.iotas.ItemStackIota
-import miyucomics.hexpose.iotas.getItemStack
 import net.minecraft.entity.ItemEntity
+import ram.talia.moreiotas.api.casting.iota.ItemStackIota
+import ram.talia.moreiotas.api.getItemStack
 
 object OpGetMaxMedia : ConstMediaAction {
 	override val argc = 1