+++ src/main/java/miyucomics/hexpose/actions/processors/OpGetVillagerData.kt
@@ -1,16 +1,19 @@
-package miyucomics.hexpose.actions.instance_data
+package miyucomics.hexpose.actions.processors
 
 import at.petrak.hexcasting.api.casting.castables.ConstMediaAction
 import at.petrak.hexcasting.api.casting.eval.CastingEnvironment
 import at.petrak.hexcasting.api.casting.getEntity
 import at.petrak.hexcasting.api.casting.iota.Iota
-import net.minecraft.entity.Entity
+import at.petrak.hexcasting.api.casting.mishaps.MishapBadEntity
+import net.minecraft.entity.passive.VillagerEntity
 
-class OpGetEntityData(private val process: (Entity) -> List<Iota>) : ConstMediaAction {
+class OpGetVillagerData(private val process: (VillagerEntity) -> List<Iota>) : ConstMediaAction {
 	override val argc = 1
 	override fun execute(args: List<Iota>, env: CastingEnvironment): List<Iota> {
 		val entity = args.getEntity(0, argc)
 		env.assertEntityInRange(entity)
+		if (entity !is VillagerEntity)
+			throw MishapBadEntity.of(entity, "villager")
 		return process(entity)
 	}
 }